requirements.txt
alembic==1.16.2
APScheduler==3.11.0
arabic-reshaper==3.0.0
argcomplete==3.6.2
blinker==1.9.0
Brotli==1.1.0
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cssselect2==0.8.0
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.58.5
greenlet==3.2.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
packaging==24.2
pdfkit==1.0.0
pillow==11.3.0
pipx==1.7.1
platformdirs==4.3.7
psycopg==3.2.9
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.2
python-bidi==0.6.6
python-dotenv==1.1.1
qrcode==8.2
reportlab==4.4.2
setuptools==78.1.0
SQLAlchemy==2.0.41
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
userpath==1.9.2
weasyprint==65.1
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.45.1
WTForms==3.2.1
zopfli==0.2.3.post1
